pyproject.toml
[project]
name = "py-django"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2",
    "django-bootstrap5>=25.1",
    "django-rosetta>=0.10.2",
    "dj-database-url>=2.3.0",
    "environs>=14.1.1",
    "gunicorn>=23.0.0",
    "uvicorn>=0.34.1",
    "whitenoise>=6.9.0",
]

[dependency-groups]
dev = [
    "django-debug-toolbar>=5.1.0",
    "ruff>=0.11.5",
    "taskipy>=1.14.1",
]

[tool.ruff]
line-length = 79
extend-exclude = ["public"]

[tool.ruff.lint]
select = ["I", "F", "E", "W", "PL", "PT"]

[tool.ruff.format]
quote-style = "single"

[tool.taskipy.tasks]
env = "cp bkp-env.txt .env"
format = "ruff check . --select I --fix && ruff format ."
migrate = "python manage.py makemigrations && python manage.py migrate"
prod = "python -m gunicorn _config.asgi:application -k uvicorn.workers.UvicornWorker"
runserver = "python manage.py runserver"
superuser = "python manage.py createsuperuser"
startapp = "python manage.py startapp"
statics = "python manage.py collectstatic --no-input"
tr = "django-admin makemessages -i .venv -i public -l"
tr-js = "django-admin makemessages -i .venv  -i public -d djangojs -l"
tr-compile = "django-admin compilemessages -i .venv -i public"